+++ test/unit_test.dart
@@ -16,6 +16,7 @@ import 'selection/selection_keyboard_test.dart' as selection_keyboard_test;
 import 'selection/selection_test.dart' as selection_test;
 import 'selection/selection_widget_test.dart' as selection_widget_test;
 import 'highlight/highlight_test.dart' as highlight_test;
+import 'render/table_scroll_test.dart' as table_scroll_test;
 import 'theme/theme_test.dart' as theme_test;
 import 'widget/render_test.dart' as render_test;
 import 'widget/widget_test.dart' as widget_test;
@@ -38,6 +39,7 @@ void main() => group('Unit', () {
       selection_widget_test.main();
       selection_keyboard_test.main();
       selection_handles_test.main();
+      table_scroll_test.main();
       render_test.main();
       widget_test.main();
     });